# services/bff_service/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Оптимизированный Gateway под frontend",
        "title": "BFF Service API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/users/api/v1/authenticate": {
            "post": {
                "description": "Этот эндпоинт проверяет данные пользователя и возвращает JWT токен.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Аутентификация пользователя по логину и паролю.",
                "parameters": [
                    {
                        "description": "Запрос аутентификации пользователя",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/users.UserAuthenticate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный ответ, содержащий JWT токен в поле 'access'",
                        "schema": {
                            "$ref": "#/definitions/users.AuthResponse"
                        }
                    },
                    "400": {
                        "description": "Ошибка валидации запроса",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка при генерации токена",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "users.AuthResponse": {
            "type": "object",
            "properties": {
                "access": {
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJleHAiOjE3MjYzOTgzOTQsInVzZXJfaWQiOjF9.1yKDMfobTV8pAOK6QFClZLRuOb44KtjoezgbBgvksBQ"
                }
            }
        },
        "users.UserAuthenticate": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
...
    }
}